+++ src/local/room_position/game_math.rs
@@ -1,14 +1,17 @@
-//! Utilities for doing math on [`LocalRoomPosition`]s which are present in the
+//! Utilities for doing math on [`Position`]s which are present in the
 //! JavaScript API.
-use crate::constants::Direction;
+use crate::{constants::Direction, objects::HasPosition};
 
-use super::LocalRoomPosition;
+use super::Position;
 
-impl LocalRoomPosition {
-    pub fn get_direction_to(self, target: LocalRoomPosition) -> Option<Direction> {
+impl Position {
+    pub fn get_direction_to<T>(self, target: &T) -> Option<Direction>
+    where
+        T: ?Sized + HasPosition,
+    {
         // Logic copied from https://github.com/screeps/engine/blob/
         // 020ba168a1fde9a8072f9f1c329d5c0be8b440d7/src/utils.js#L73-L107
-        let (dx, dy) = self - target;
+        let (dx, dy) = self - target.pos();
         if dx.abs() > dy.abs() * 2 {
             if dx > 0 {
                 Some(Direction::Right)
@@ -37,27 +40,40 @@ impl LocalRoomPosition {
     }
 
     #[inline]
-    pub fn get_range_to(self, target: LocalRoomPosition) -> u32 {
-        let (dx, dy) = self - target;
+    pub fn get_range_to<T>(self, target: &T) -> u32
+    where
+        T: ?Sized + HasPosition,
+    {
+        let (dx, dy) = self - target.pos();
         dx.abs().max(dy.abs()) as u32
     }
 
     #[inline]
-    pub fn in_range_to(self, target: LocalRoomPosition, range: u32) -> bool {
+    pub fn in_range_to<T>(self, target: &T, range: u32) -> bool
+    where
+        T: ?Sized + HasPosition,
+    {
         self.get_range_to(target) < range
     }
 
     #[inline]
-    pub fn is_equal_to(self, target: LocalRoomPosition) -> bool {
-        self == target
+    pub fn is_equal_to<T>(self, target: &T) -> bool
+    where
+        T: ?Sized + HasPosition,
+    {
+        self == target.pos()
     }
 
     /// True if this position is in the same room as the target, and the range
     /// is at most 1.
     #[inline]
-    pub fn is_near_to(self, target: LocalRoomPosition) -> bool {
-        self.room_name() == target.room_name()
-            && (self.x() as i32 - target.x() as i32).abs() <= 1
-            && (self.y() as i32 - target.y() as i32).abs() <= 1
+    pub fn is_near_to<T>(self, target: &T) -> bool
+    where
+        T: ?Sized + HasPosition,
+    {
+        let pos = target.pos();
+        self.room_name() == pos.room_name()
+            && (self.x() as i32 - pos.x() as i32).abs() <= 1
+            && (self.y() as i32 - pos.y() as i32).abs() <= 1
     }
 }