+++ src/objects.rs
@@ -18,6 +18,7 @@ use stdweb_derive::ReferenceType;
 
 use crate::{
     constants::{ResourceType, ReturnCode, StructureType},
+    local::Position,
     macros::*,
     traits::{IntoExpectedType, TryFrom, TryInto},
     ConversionError,
@@ -55,8 +56,6 @@ reference_wrappers!(
     Room,
     #[reference(instance_of = "RoomObject")]
     RoomObject,
-    #[reference(instance_of = "RoomPosition")]
-    RoomPosition,
     #[reference(instance_of = "Room.Terrain")]
     RoomTerrain,
     #[reference(instance_of = "Source")]
@@ -114,11 +113,11 @@ reference_wrappers!(
 ///
 /// This can be freely implemented for anything with a way to get a position.
 pub trait HasPosition {
-    fn pos(&self) -> RoomPosition;
+    fn pos(&self) -> Position;
 }
 
-impl HasPosition for RoomPosition {
-    fn pos(&self) -> RoomPosition {
+impl HasPosition for Position {
+    fn pos(&self) -> Position {
         self.clone()
     }
 }
@@ -128,8 +127,8 @@ impl<T> HasPosition for T
 where
     T: RoomObjectProperties,
 {
-    fn pos(&self) -> RoomPosition {
-        js_unwrap_ref!(@{self.as_ref()}.pos)
+    fn pos(&self) -> Position {
+        Position::from_packed(js_unwrap!(@{self.as_ref()}.pos.__packedPos))
     }
 }