+++ src/objects/impls/flag.rs
@@ -16,7 +16,7 @@ simple_accessors! {
 
 impl Flag {
     /// Useful method for constructing Flag from the result of
-    /// `RoomPosition.createFlag` or `Room.createFlag`.
+    /// `Position.createFlag` or `Room.createFlag`.
     ///
     /// String names are mapped to Ok(Ok(s)), return codes are mapped to
     /// Ok(Err(e)), other unknown inputs are mapped to Err(e).
@@ -50,9 +50,9 @@ impl Flag {
     }
 
     pub fn set_position<T: HasPosition>(&self, pos: T) {
-        let room_pos = pos.pos();
+        let pos = pos.pos();
         js! { @(no_return)
-            @{self.as_ref()}.setPosition(@{room_pos.as_ref()});
+            @{self.as_ref()}.setPosition(pos_from_packed(@{pos.packed_repr()}));
         }
     }